pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "coriro"
dynamic = ["version"]
description = "Color measurement instrument for vision-language model (VLM) pipelines"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
    { name = "Saad Irfan", email = "saad@coriro.org" },
]
keywords = [
    "color",
    "extraction",
    "vlm",
    "vision-language-model",
    "oklch",
    "palette",
    "sidecar",
    "measurement",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.20.0",
    "Pillow>=8.0.0",
]

[project.optional-dependencies]
text = ["pytesseract>=0.3.0"]
accents = ["scipy>=1.7.0"]
cnn = ["torch>=2.0.0", "timm>=0.9.0"]
all = [
    "pytesseract>=0.3.0",
    "scipy>=1.7.0",
    "torch>=2.0.0",
    "timm>=0.9.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.urls]
Homepage = "https://coriro.org"
Documentation = "https://coriro.org/docs"
"Source Code" = "https://github.com/coriro-org/coriro"
"Bug Tracker" = "https://github.com/coriro-org/coriro/issues"

[tool.setuptools.dynamic]
version = { attr = "coriro.__version__" }

[tool.setuptools.packages.find]
include = ["coriro*"]
exclude = ["tests*", "coriro.adapters*", "coriro.benchmark*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.package-data]
coriro = ["py.typed"]
